Moving html skeleton generation code to ftplugin
This commit illustrates how a specific command can be executed when a
new file of a specific filetype is created. This is opposed to the
autocmd lines used in my vimrc, which match a certain _extension_. Here,
I match a specific filetype.

The problem and its solution are well explained in the following SO
post: https://stackoverflow.com/questions/23496243/autocommand-for-new-file-of-specific-filetype